Add Styles.swift and better naming
