--- FreeOTP/UIFont+Ext.swift
@@ -1,32 +1,0 @@
-//
-// FreeOTP
-//
-// Authors: Nathaniel McCallum <npmccallum@redhat.com>
-//
-// Copyright (C) 2015  Nathaniel McCallum, Red Hat
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-
-import UIKit
-
-extension UIFont {
-    static func dynamicSystemFont(ofSize fontSize: CGFloat, weight: UIFont.Weight) -> UIFont {
-        if Screen.size == .large {
-            return .systemFont(ofSize: fontSize + 4, weight: weight)
-        } else if Screen.size == .medium {
-            return .systemFont(ofSize: fontSize + 2, weight: weight)
-        }
-        return .systemFont(ofSize: fontSize, weight: weight)
-    }
-}
